pyproject.toml
[project]
name = "rime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "audioop-lts>=0.2.1",
    "certifi>=2024.0.0",
    "deepgram-sdk>=6.0.1",
    "fastapi>=0.135.1",
    "google-genai>=1.0.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "miniaudio>=1.61",
    "openai>=2.24.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.15.4",
    "twilio>=9.10.2",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
]
